requirements.txt
absl-py==0.1.10
cupy-cuda111==8.1.0 
chainer==7.8.1
Cython==0.27.3
fastrlock==0.8
h5py
meshzoo
networkx==2.6.3
numpy==1.21.5
packaging==21.3
Pillow==9.0.0
progressbar==2.5
protobuf==3.9.2
PyOpenGL==3.1.0
pyparsing==3.0.7
PyWavelets==1.2.0
scikit-image==0.19.1
scipy==1.7.3
tifffile==2021.11.2
torch==1.5.0
tqdm==4.62.3
visdom==0.1.7
torchvision==0.6.0
chumpy==0.70
imageio==2.15.0
ipython==7.31.1
jsonpickle 
humanfriendly 
matplotlib==3.5.1
opencv-python==4.6.0.66
pandas==1.3.5
progressbar2==3.51.0
pretrainedmodels==0.7.4
torchnet==0.0.4
tensorflow==2.8.0
requests
scikit-image==0.19.1
statistics 
